Accept float frequencies in biquad filter config and SET_TMC_BIQUAD_FILTER

+++ tmc4671_biquad.py
@@ -9,7 +9,7 @@ from typing import NamedTuple
 
 class BiquadFilter(NamedTuple):
     type: str
-    freq: int
+    freq: float
     slope: float
